requirements.txt
streamlit
streamlit_chat
langchain
openai
python-dotenv
tiktoken
chromadb
apify-client
